remove duplicated files and fix id types relation between entity and repository

+++ src/main/java/com/dovendev/track/jpa/repositories/TrackLinkRepository.java
@@ -4,6 +4,6 @@ import com.dovendev.track.jpa.entities.TracksResources;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface TrackLinkRepository extends JpaRepository<TracksResources, Integer> {
+public interface TrackLinkRepository extends JpaRepository<TracksResources, Long> {
 }
